Inline MessageHandler in MessageWorker
